+++ changelog.md
@@ -3,3 +3,7 @@
 ## 1.0.0
 
 * Initial release
+
+## Next
+
+* Build script to copy tests from `co-series`